Refactor commit generation logic into CommitManager class
- Move commit-related functionality from commit.ts to new CommitManager class
- Update commit.ts to use CommitManager for handling commit operations
- Improve code organization and maintainability
